+++ wsei-todo-app/src/constans/Colors.ts
@@ -3,6 +3,5 @@ export default {
   black: "#000",
   red: "#fc0303",
   orange: "#FFCC66",
-  lightBlue: "#51CBE9",
-  darkerWhite: "#cccccc",
+  background: "#0099cc",
 };