+++ wsei-todo-app/src/screens/Home/index.tsx
@@ -6,20 +6,33 @@ import { useNavigation } from "@react-navigation/native";
 import Colors from "../../constans/Colors";
 
 const WelcomeText = styled.Text`
-  margin: 80px 20px;
-  font-size: 30px;
+  font-size: 35px;
+  flex: 1;
+  margin-top: auto;
+  margin-bottom: auto;
   color: ${Colors.black};
+  text-align: center;
 `;
 const Textarea = styled.Text`
   margin: 10px 20px;
   font-size: 20px;
   color: ${Colors.black};
 `;
-
 const CustomImage = styled.Image`
   width: 160px;
   height: 160px;
-  margin-top: 30px;
+
+  border-radius: 100px;
+  margin-top: 40px;
+`;
+const Header = styled.View`
+  height: 250px;
+  padding: 10px;
+  margin-bottom: 30px;
+  flex-direction: row;
+`;
+const Body = styled.View`
+  background-color: ${Colors.background};
 `;
 
 interface IHomeProps {}
@@ -28,21 +41,12 @@ const Home: FC<IHomeProps> = (props) => {
   const navigation = useNavigation();
 
   return (
-    <View
-      style={{
-        backgroundColor: Colors.lightBlue,
-        flexDirection: "row",
-        justifyContent: "space-between",
-      }}
-    >
+    <Body>
       <ScrollView>
-        <CustomImage
-          style={{
-            borderBottomRightRadius: 20,
-          }}
-          source={require("../../assets/logo.png")}
-        />
-        <WelcomeText>Welcome in codelife!!!</WelcomeText>
+        <Header>
+          <CustomImage source={require("../../assets/logo.png")} />
+          <WelcomeText>Welcome in codelife!!!</WelcomeText>
+        </Header>
         <Textarea>
           Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quod sed quo
           qui adipisci commodi odio ut deserunt deleniti dolore unde nostrum
@@ -60,7 +64,7 @@ const Home: FC<IHomeProps> = (props) => {
           vitae voluptates veritatis.
         </Textarea>
       </ScrollView>
-    </View>
+    </Body>
   );
 };