+++ wsei-todo-app/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -8,8 +9,67 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-    - 
+    -
 
 ### Added
 
     - Initial app config
+    ## [1.0.1] - 2020-03-01
+
+## [1.0.1] - 2020-03-02
+
+### Changes
+
+    -BottomNavigation.tsx
+
+### Added
+
+    -Bottom bar
+
+## [1.0.2] - 2020-03-05
+
+### Changes
+
+    -Welcome/index.tsx
+
+### Added
+
+    -Home Page
+    -TodoList Page
+    -Info Page
+
+## [1.0.3] - 2020-03-30
+
+### Changes
+
+### Added
+
+    -SplashScreen
+
+## [1.1.0] - 2020-05-05
+
+### Changes
+
+    -ListModal/index.tsx
+    -RemoveListButton.tsx
+
+### Added
+
+    -Commit/Push Project on GitHub
+
+## [1.2.0] - 2020-05-22
+
+### Changes
+
+    -ListModal/index.tsx
+
+### Added
+
+    -Adding tasks
+    -Button Back
+
+## [1.2.0] - 2020-05-27
+
+### Changes
+
+    -Styling CSS